Basic functional app implemented
